+++ src/managednetworkfabric/HISTORY.rst
@@ -9,6 +9,8 @@ Release History
 * az core cli updated to version 2.67, aaz_dev updated to version 3.2.0, and azdev to version 0.1.94.
 * New `fabric identity` commands introduced: `fabric identity assign`, `fabric identity remove`, `fabric identity show`.
 * Expose the 'update' command for the 'interface' resource that was previously removed in post generate script.
+* Making the Network Monitor resource available through the CLI.
+* Within the the IPv4/IPv6 PrefixLimitPatchProperties structure for External and Internal Networks, renamed the 'softLimit' property to 'threshold'.
 
 7.1.0
 ++++++